descw-1283 updates model and template
